requirements.txt
alembic==1.14.1
bcrypt==4.2.1
blinker==1.9.0
cachelib==0.9.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
Flask==3.1.0
Flask-Caching==2.3.0
Flask-Migrate==4.1.0
flask-redis==0.4.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.9
MarkupSafe==3.0.2
psycopg2==2.9.10
pycparser==2.22
PyMySQL==1.1.1
python-dotenv==1.0.1
redis==5.2.1
SQLAlchemy==2.0.38
typing_extensions==4.12.2
Werkzeug==3.1.3
gunicorn
requests~=2.32.3
razorpay~=1.4.2
authlib